feat: replace branch icon with dynamic agent status indicator
Show colored circles in place of the branch icon when an agent is
active — orange for running, red for needs_attention, branch icon
when idle.
